pyproject.toml
[tool.poetry]
name = "vehicle-dynamics-planner"
version = "2025.01"
description = "Sampling-based planner based on analytical vehicle models,samples are trained with neural networks."
license = "LGPL-3.0-or-later"
authors = [
    "Marc Kaufeld <marc.kaufeld@tum.de>",
]
maintainers = [
    "Marc Kaufeld <marc.kaufeld@tum.de>",
]
readme = "README.md"

packages = [
    { include = "scripts" },
    { include = "ml_planner"},
]

[tool.poetry.dependencies]

python = ">=3.9,<3.12"

commonroad-vehicle-models = "^3.0.2"
commonroad-io = "~2024.3"
commonroad-route-planner = "~2025.0.1"
commonroad-velocity-planner = "^0.1.1"
commonroad-drivability-checker = "~2024.1"
wale-net = "~3.0.3"
casadi = "^3.7.0"
bokeh = "^3.5.0"
hydra-core = "^1.3.2"
control = "^0.10.0"
matplotlib = "^3.8.0"
tqdm = "^4.67.0"
torch = "^2.5.1"
tensorboard = "^2.18.0"
pandas = "^2.2.3"
imageio = "^2.31.2"

# frenetix-motion-planner = { git = "https://gitlab.lrz.de/av2.0/commonroad/commonroad-reactive-planner.git", branch="master_frenetix"}

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
sphinx = ">=6.2.1"

[tool.poetry.group.dev]
optional = true

[tool.poetry.dev-dependencies]
pytest = "^8.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.5.1"

# [tool.poetry.scripts]
# reactive-planner = "main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
